Updated common utility class
