+++ appserver/admingui/common/src/main/java/org/glassfish/admingui/common/security/AdminConsoleAuthModule.java
@@ -70,6 +70,8 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import static com.sun.logging.LogCleanerUtil.neutralizeForLog;
+
+import org.glassfish.common.util.InputValidationUtil;
 import org.glassfish.jersey.client.authentication.HttpAuthenticationFeature;
 
 import org.glassfish.hk2.api.ServiceLocator;
@@ -294,11 +296,11 @@ public class AdminConsoleAuthModule implements ServerAuthModule {
                     origRequest = "/index.jsf";
                 }
                 logger.log(Level.INFO, "Redirecting to {0}", neutralizeForLog(origRequest));
-                if (GuiUtil.validateStringforCRLF(origRequest)) {
+                if (InputValidationUtil.validateStringforCRLF(origRequest)) {
                     response.sendError(403,"Forbidden");
                 }
                 response.sendRedirect(
-                        response.encodeRedirectURL(GuiUtil.removeLinearWhiteSpaces(origRequest)));
+                        response.encodeRedirectURL(InputValidationUtil.removeLinearWhiteSpaces(origRequest)));
             } catch (Exception ex) {
                 AuthException ae = new AuthException();
                 ae.initCause(ex);
