+++ appserver/web/web-core/src/main/java/org/apache/catalina/connector/Response.java
@@ -88,7 +88,7 @@ import org.apache.catalina.core.StandardContext;
 import org.apache.catalina.security.SecurityUtil;
 import org.apache.catalina.util.CharsetMapper;
 import org.apache.catalina.util.RequestUtil;
-import org.apache.catalina.util.ResponseUtil;
+import org.glassfish.common.util.InputValidationUtil;
 import org.glassfish.grizzly.http.util.CharChunk;
 import org.glassfish.grizzly.http.util.CookieSerializerUtils;
 import org.glassfish.grizzly.http.util.CookieUtils;
@@ -1542,8 +1542,8 @@ public class Response
             return;
 
         try {
-            String safeName = ResponseUtil.getSafeHeaderName(name);
-            String safeValue = ResponseUtil.getSafeHeaderValue(value);
+            String safeName = InputValidationUtil.getSafeHeaderName(name);
+            String safeValue = InputValidationUtil.getSafeHeaderValue(value);
             coyoteResponse.setHeader(safeName, safeValue);
         } catch (Exception e) {
             try {
