+++ appserver/web/web-core/src/main/java/org/apache/catalina/util/ResponseUtil.java
@@ -59,20 +59,12 @@
 package org.apache.catalina.util;
 
 import javax.servlet.ServletOutputStream;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.PrintWriter;
 import java.io.Reader;
-import java.util.regex.Pattern;
 
 public final class ResponseUtil {
-
-    public static final String CRLF_ENCODED_STRING_LOWER = "%0d%0a";
-    public static final String CRLF_ENCODED_STRING_UPPER = "%0D%0A";
-    public static final String CR_ENCODED_STRING_LOWER = "%0d";
-    public static final String CR_ENCODED_STRING_UPPER = "%0D";
-    public static final String CRLF_STRING = "\"\\r\\n\"";
 
     /**
      * Copies the contents of the specified input stream to the specified
@@ -135,78 +127,4 @@ public final class ResponseUtil {
         return exception;
 
     }
-
-    /**
-     Validate the String for Header Injection Attack.
-
-     @param input	String to be validate
-     @return		boolean
-     */
-    public static boolean validateStringforCRLF (String input) {
-        if (input != null && (input.contains(CRLF_ENCODED_STRING_LOWER)
-                || input.contains(CRLF_ENCODED_STRING_UPPER)
-                || input.contains(CR_ENCODED_STRING_UPPER)
-                || input.contains(CR_ENCODED_STRING_LOWER)
-                || input.contains(CRLF_STRING))) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     * Remove unwanted white spaces in the URL.
-     *
-     * @param input	String to be stripped with whitespaces
-     * @return		String
-     */
-    public static String removeLinearWhiteSpaces(String input) {
-        if (input != null) {
-            input = Pattern.compile("//s").matcher(input).replaceAll(" ");
-        }
-        return input;
-    }
-
-    /**
-     * Return Http Header Name after suitable validation
-     *
-     * @param headerName Header Name which should be validated before being set
-     * @return String Header Name sanitized for CRLF attack
-     */
-    public static String getSafeHeaderName(String headerName) throws Exception {
-        headerName = removeLinearWhiteSpaces(headerName);
-            if (validateStringforCRLF(headerName)) {
-                throw new Exception("Header Name invalid characters");
-            }
-        return headerName;
-    }
-
-    /**
-     * Return Http Header Value after suitable validation
-     *
-     * @param headerValue Header Value which should be validated before being set
-     * @return String Header Value sanitized for CRLF attack
-     */
-    public static String getSafeHeaderValue(String headerValue) throws Exception {
-        headerValue = removeLinearWhiteSpaces(headerValue);
-        if (validateStringforCRLF(headerValue)) {
-                throw new Exception("Header Value invalid characters");
-            }
-        return headerValue;
-    }
-
-    /**
-     * Return Cookie Http Header Value after suitable validation
-     *
-     * @param headerValue Header Value which should be validated before being set
-     * @return String Header Value sanitized for CRLF attack
-     */
-    public static String getSafeCookieHeaderValue(String headerValue) throws Exception {
-        headerValue = removeLinearWhiteSpaces(headerValue);
-        if (validateStringforCRLF(headerValue)) {
-            throw new Exception (" Cookie Header Value has invalid characters");
-        }
-        return headerValue;
-    }
-
 }
