+++ appserver/web/web-glue/src/main/java/com/sun/enterprise/web/VirtualServerPipeline.java
@@ -58,8 +58,8 @@ import java.util.ResourceBundle;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.util.regex.Pattern;
 
+import org.glassfish.common.util.InputValidationUtil;
 import org.glassfish.grizzly.http.util.CharChunk;
 import org.glassfish.web.LogFacade;
 
@@ -84,12 +84,6 @@ public class VirtualServerPipeline extends StandardPipeline {
     private ArrayList<RedirectParameters> redirects;
 
     private ConcurrentLinkedQueue<CharChunk> locations;
-
-    public static final String CRLF_ENCODED_STRING_LOWER = "%0d%0a";
-    public static final String CRLF_ENCODED_STRING_UPPER = "%0D%0A";
-    public static final String CR_ENCODED_STRING_LOWER = "%0d";
-    public static final String CR_ENCODED_STRING_UPPER = "%0D";
-    public static final String CRLF_STRING = "\"\\r\\n\"";
 
     /**
      * Constructor.
@@ -321,40 +315,15 @@ public class VirtualServerPipeline extends StandardPipeline {
             }
 
             // Validate the URL for extra spaces before redirection.
-            if(validateStringforCRLF(location)) {
+            if(InputValidationUtil.validateStringforCRLF(location)) {
                 hres.sendError(403, "Forbidden");
             } else {
-                hres.sendRedirect(removeLinearWhiteSpaces(location));
+                hres.sendRedirect(InputValidationUtil.removeLinearWhiteSpaces(location));
             }
             return true;
         }
 
         return false;
-    }
-
-    public static boolean validateStringforCRLF (String input) {
-        if (input != null && (input.contains(CRLF_ENCODED_STRING_LOWER)
-                || input.contains(CRLF_ENCODED_STRING_UPPER)
-                || input.contains(CR_ENCODED_STRING_UPPER)
-                || input.contains(CR_ENCODED_STRING_LOWER)
-                || input.contains(CRLF_STRING))) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     Remove extra white spaces in String and convert into lowercase.
-
-     @param input Input String
-     @return		string
-     */
-    public static String removeLinearWhiteSpaces(String input) {
-        if (input != null) {
-            input = Pattern.compile("//s").matcher(input).replaceAll(" ");
-        }
-        return input;
     }
 
     /**
